Crate lispdm

LispDM is a Scheme interpreter. Currently, it supports only a subset of Scheme.

use lispdm::Engine;
let mut engine = Engine::default();

let result = engine.eval::<i64>("(+ 1 2)").unwrap();
assert_eq!(result, Ok(3));

Macros

exprs

A helper macro to construct Exprs from given expressions

Structs

Engine

Provides the main functionality of LispDM.

EnvRef

Reference to the environment.

InputPort

Input port.

List

List of expressions.

OutputPort

Output port.

Enums

Arity

Represents the arity of a procedure.

Expr

Represents all possible values in interpreter.

LispDMError

Error type that represents all possible errors in LispDM.

Port

Port type that can be either input or output.

ProcedureKind

Represents the kind of a procedure.

ProcedureReturn

The return value of a procedure. Used to determine whether a tail call should be performed.

Constants

PRELUDE

Prelude of LispDM.

Traits

FromExpr

The exit point for turning Expr into Rust types.

Type Aliases

Exprs

A list of Exprs. Also can be created with exprs! macro.

FromExprResult

Represents result of conversion from Expr to another type

ProcedureFn

The type of a procedure function.

ProcedureResult

The result of a procedure call.