Enum lispdm::Expr

pub enum Expr {
    Integer(i64),
    Float(f64),
    Symbol(String),
    String(Rc<RefCell<String>>),
    Char(char),
    Boolean(bool),
    List(List),
    Void,
    Procedure(Procedure),
    InputPort(Rc<RefCell<dyn InputPortSuperTrait>>),
    OutputPort(Rc<RefCell<dyn OutputPortSuperTrait>>),
}

Represents all possible values in interpreter.

This is the only type that can be used in interpreter. All other types are converted to this one.

Variants

Integer(i64)

Integer number

Float(f64)

Real number

Symbol(String)

Symbol

String(Rc<RefCell<String>>)

A reference to mutable string.

Char(char)

Character

Boolean(bool)

Boolean

List(List)

An immutable list of expressions.

'() (null) is represented as empty list. Dotted lists (nested pairs) and proper lists are represented as list of expressions.

Internal List struct does not exported, so you can’t construct list directly. To construct a list you should use Expr::new_empty_list, Expr::new_proper_list and Expr::new_dotted_list instead.

NOTE: Since lists are immutable, they are flattened on creation. This means that (1 2 3) and (1 . (2 . (3 . '())) are represented as same list internally.

Void

Unspecified value

This expression is created only on evaluation stage and evaluating it will cause error.

Procedure(Procedure)

Procedure

This expression is created only on evaluation stage and evaluating it will cause error.
Because of this, internal Procedure struct does not exported, so you can’t construct procedure directly.

InputPort(Rc<RefCell<dyn InputPortSuperTrait>>)

Input port

OutputPort(Rc<RefCell<dyn OutputPortSuperTrait>>)

Output port

Implementations

impl Expr

pub fn new_empty_list() -> Self

Creates new empty list '()

pub fn new_dotted_list(list: Exprs) -> Self

Creates new dotted list like '(1 2 . 3) from Exprs

Last element of list is used as cdr of last nested pair

If list is empty, returns empty list

pub fn new_proper_list(list: Exprs) -> Self

Creates new proper list like '(1 2 3) from Exprs

pub fn new_string<S: Into<String>>(string: S) -> Self

Creates new Expr::String from any type that implements Into<String>

pub fn new_symbol<S: Into<String>>(string: S) -> Self

Creates new Expr::Symbol from any type that implements Into<String>

pub fn kind(&self) -> &'static str

Returns string representation of the type of self

pub fn is_truthy(&self) -> bool

Returns true if self represents a true value

pub fn is_symbol(&self) -> bool

Checks if self is a Expr::Symbol

pub fn is_specific_symbol(&self, name: &str) -> bool

Checks if self is a Expr::Symbol with given name

pub fn is_list(&self) -> bool

checks if self is a Expr::List

pub fn is_empty_list(&self) -> bool

Checks if self represents an empty list '()

pub fn is_proper_list(&self) -> bool

Checks if self represents a proper list

pub fn is_dotted_list(&self) -> bool

Checks if self represents a dotted list This means that last element of last nested list is not '()

pub fn is_string(&self) -> bool

Checks if self is a Expr::String

pub fn is_char(&self) -> bool

Checks if self is a Expr::Char

pub fn is_integer(&self) -> bool

Checks if self is a Expr::Integer

pub fn is_float(&self) -> bool

Checks if self is a Expr::Float

pub fn is_boolean(&self) -> bool

Checks if self is a Expr::Boolean

pub fn is_void(&self) -> bool

Checks if self is a Expr::Void

pub fn is_procedure(&self) -> bool

Checks if self is a Expr::Procedure

pub fn is_input_port(&self) -> bool

Checks if self is a Expr::InputPort

pub fn is_output_port(&self) -> bool

Checks if self is a Expr::OutputPort

pub fn into<T: FromExpr>(self) -> FromExprResult<T>

Tries to convert self into T, which is a type that implements FromExpr