Trait lispdm::FromExpr

pub trait FromExpr: Sized {
    // Required method
    fn from_expr(expr: Expr) -> FromExprResult<Self>;
}

The exit point for turning Expr into Rust types.

This trait implemented for most primitives. You can also manually implement this for any type.

Implemented conversions

Expr	Rust type

Expr::Void	unit

Expr::Boolean	bool

Expr::Char	char

Expr::Integer	i64

Expr::Float	f64

Expr::Symbol	String

Expr::String	Rc<RefCell<String>>

proper Expr::List	Vec<T> or VecDeque<T>, where T is a type that implements FromExpr

dotted Expr::List with 2 elements	(A, B), where A and B are types that implements FromExpr

Required Methods

fn from_expr(expr: Expr) -> FromExprResult<Self>

Tries to convert Expr into Self. Returns FromExprResult<Self> with result of conversion.

Object Safety

This trait is not object safe.